refactor: signup button copy experiment
We want to try out different copies for the login button in the header.
For this purpose I added a new feature `signup_button_copy` in Flagsmith. This feature will hold the string for the button.
Please note, in mobile resolution we will keep the current copy due to space restrictions.

I also migrated the relevant signup events to our in-house analytics tracking so we can monitor the performance of the experiment.

DD-362 #done