fix: update alerts function was not defined in the alerts context
Due to our recent change in the boot provider, we mistakenly renamed the `updateAlerts` function.
It caused critical errors throughout the app when trying to call this function as it was undefined.
The most common case is when the feed filters had the alert on and the app need to turn off this alert by dispatching the call.
Key learning here is to use explicit types in functions like `useMemo` and create test cases to prevent it.