fix: registration stuck-up on logout (#777)
When logging out on the registration page, it gets stuck because we reload when the user logs out. So now, we either redirect when the redirect_uri is available on the URL query params or the user is at the registration page.

With that we still kept the reload for all other cases.