fix: redirect user to registration if not yet confirmed (#778)
When the user has not confirmed his info, we should redirect them to fill up the form. This fixes the bug for when the user clicked the dailydev logo it should redirect them back to finish the form.

There was also an issue that when you cancel the registration the user gets stuck on an empty screen. This was due to reloading when logging out, which in registration - as an empty page should redirect to where the user has started.