feat: my feed (#783)
DD-354 #done

We wanted to emphasize the relevant articles shown by configuring the feed filters. With it, we want to make the changes much easier and noticeable by applying different ways to attract in making the adjustments.

This is, for the moment, an experiment to see how much would be the difference between users we can convert and people who will be having their filters configured than before. If successful, we will be running multiple experiments and it will open the door for creating custom feeds per user.