fix: track active page state correctly
Our feed engine uses the client indication of whether the session is active or not for certain decisions.
After some data analysis and debugging, I found out that there's some issue in how we track it.
The edge case open new tab -> type url -> navigate to the new url will be marked as an active session.
This fix makes sure the user clicked somewhere in the window object or at least scrolled down a bit before the session marked as active.
I also fixed an edge case where events were not sent on page unload.