refactor: opt-in settings for utilizing ranking algorithm by sorting date  (#822)
DD-388 #done

Due to users' request for the recent feed to be brought back, we have decided to have something that provides the same situation as it was realized the recent feed is not actually a feed but an algorithm for how we fetch the articles.

With that, we have decided to allow users to have an opt-in setup to allow changes on how we based the ranking of articles. This PR should provide a dropdown for selecting which ranking algorithm to use and have a place in customizing the settings.