feat: custom links (#843)
DD-85 #done

As part of being a browser extension, the ability to navigate to quick links provided by the browser itself has been overridden. We have provided an alternative in the past but the user felt custom links would be a great addition. This now allows anonymous users or registered users to fully manage their quick links.