fix: page container placement (#855)
DD-435 #done
DD-385 #done

The content push was due to race in render where a side-by-side component is displayed asynchronously, which results in shifting the current content. As discussed we have now centered our content in the absolute screen which reduces the jumpyness when toggling the sidebar. We have now defined a specific PageContainer to be utilized by existing pages and future ones.