fix: profile tooltip rank version (#905)
Profile Tooltip was still using the v1 of our reading rank which leads to confusion as the profile reflects the correct value. We now point it to utilize v2 and we have removed a package that hasn't been used anymore. Then a fix for the missing space in the tooltip label for the profile's heatmap.