fix: dnd banner missing when active (#912)
DD-563 #done

The DnD's banner is missing whenever enabled, this was due to the changes we applied for the re-layout which we utilize two separate components for banners. We are now allowing our main layout component to accept custom banners to render or if empty, will default to our promotional banner.