fix: do not show new rank on opt-out gamification (#907)
DD-542 #done

We now have an opt-out toggle for the weekly goal and when you try to toggle it on and off, it will suddenly display the new rank popup modal whenever you are eligible. We now restrict it based if the settings modal is opened to ensure the correctness of when to actually display it.