fix: flaky tests for rank-related (#914)
DD-565 #done

The flaky test was the result of an undefined value for the cached state which gets updated all the time as an anonymous user due to not having checked if the user context cache has loaded yet. Now we will wait until we load from the cache to ensure the correct condition had passed based on the actual requirements.