feat: add device id to analytics events
Device ID is used to identify the specific device that is sending the events to our server.
It is used to filter out bots, cookie blockers, and other users who might generate noise in our tracking system.
It is generated by the client and saved locally in WebSQL so it doesn't rely on cookies.