refactor: search field to use hook for debounce (#922)
DD-569 #done

We have a utility hook for debouncing, but we also need to consider passing a parameter for certain cases. We have refactored it to allow arguments to be passed and be typescript to ensure the parameters sent are right. The primary concern in this PR though is to apply the debouncing on our search field for posts.