feat: mention user in comments (#906)
DD-469 #done

We wanted our users to be able to mention other users in the comments section. We created a custom hook to manage all the complexities that came with it. Once anyone mentioned a user, it should render differently and must have a highlight based on our design spec.