fix: at button borderless and on comment hover (#935)
There was an issue when a user hovered on a mention and went out. Once the user goes to hover on any part of the comment box where it was rendered, it would show the tooltip. We now explicitly manage the state for when should the tooltip be visible. This PR also removed the at button's border.