fix: article modal styling - simplified stylings (#961)
There were a lot of things going on with the StyledModal CSS, which led to complexities. We tried utilizing it before but it became apparent that it is very difficult to maintain and support different browsers.

With the current approach, it is much more simplistic and requires no recently introduced properties like fit-content, etc.