refactor: post page to include header for post options (#969)
The post options menu was missing on the actual post page due to the major change in the layout. We now allow the post navigation to be displayed whether there are left and right actions. The mobile screen size of our post page should now be almost the same as the post modal without only the left and right buttons.