fix: upvote modal user list profile tooltip (#957)
The upvote popup modal contains a list of users and for any user displaying a component, we allow the users to know more about them by hovering. It should now allow users to see the tooltip on popup modals. We also applied some mechanism to hide the popup when the user has made any scrolling.