refactor: comment box from textarea to div contenteditable (#959)
The comment box initially was already a div with content editable property to true but while working with the user mention feature, it seemed much easier to accomplish with a textarea. Though the negative side was the height of the input element is not reactive anymore.

We've decided to support the user mention while using div with content editable, then we took the implementations of textarea to guide us in a much simpler solution with the new one.