fix: companion missing at some point (#1005)
WT-120 #done

URLs with a target element (i.e. #container), it won't be able to find the actual post relevant to the actual link. So now we strip those parts. Probable cause of why companion goes missing at some point is that users have probably clicked a header or clicked a ToC item that led them to navigate the same page but different url.