fix: layout shift after making a search (#1048)
WT-166 #done

After selecting an option in the search, the suggestion menu is taking so much space in mobile view that makes it overflow horizontally. We now set a specific value (0) whenever the suggestion is not visible but in DOM. This also fixes the issue when you have pressed enter and typed another query, the suggestions won't show.