refactor: feed post card layout (#1016)
WT-51 #done

Multiple AB test will be introduced on two different versions of our feed cards to determine whether opening a modal on post card click is much more desirable for the users or not. With that, will allow the users to have a button to directly visit the actual article's page. We also introduced a custom context for feature flag-related variables for easier management in development.