fix: companion popup placement (#1070)
WT-198 #done

The companion popup's placement is a bit unstable, after further checking, the tooltip library has a default max-width which produces an issue with how it calculates the actual width for the correct placement. We have removed the max-width value from tippyjs as it doesn't seem to be something important on their end.