fix: post context menu ids (#1124)
WT-297 #done

The context menu for the post reporting posts was reusable, though the IDs were not included as a parameter, hence, every instance is having the same ID and is triggered by the same call for all other instances. We now require an additional props to include context menu id.