refactor: styled modal class selector (#1193)
WT-437 #done

Our modals conflict with one another when it comes to being displayed on top of another. This was due to using the same class `:global` selector. To make it easier to override and avoid using the base styles, we now utilize tailwindcss classes since it will have weaker class selector and make CSS module overriding work as intended.