fix: applying new headline button guideline (#1347)
WT-539 #done

The place where we display the button to open the feed filters menu was changed, and it was decided to create a feed heading to be a button in such cases. "My feed" heading will become a button once the user has decided to play around with their filters.