fix: user mention on mobile issue (#1591)
WT-1031

This fixes the issue where comment mention is not working on android, the root cause was, in Android, keydown or keyup events are not recognized, hence, the functions were not getting triggered.

The listener that works across platforms is oninput but they have some differences (keyboard events are not recognized, since there is no change on the input). I have introduced the enums to identify whether the event was a KeyboardEvent such as navigating, etc. to properly execute the necessary functions.