feat: indicate when comment author has no access (#1722)
Request a query to check the user's membership.
The type for the comments area is a bit messy, so I've taken the liberty to do some code cleanups. There were a lot of spread properties where some are used in other places and some are not. Taking the advice Ante has suggested in the past, I have introduced a single param for the whole object.

WT-1202 #done