feat: growthbook support (#1958)
Add support in our new experimentation platform. This is done using their react package with a bit of sugar coating on top to optimize performance using our boot endpoint. Feature rules are encrypted which to prevent from users to understand our experimentation rules. This is also a concern as I'm not sure how decrypting will impact our performance.

Another thing to consider, we have to check the additional bundle size. Perhaps we should take only the parts we use from the growthbook package because there's so many features there but we barely use them.