fix: evaluate edges length for comments count
as noted in https://github.com/dailydotdev/apps/pull/2101#issuecomment-1695410789 feed post cache

is not updated correctly so numComments stays 0 even though user commented

this created an edge case where if post had no comments it would always show "be first to comment"

placeholder in modal since numComments would stay the same until post is refetched.

this evaluates postComments edges to get the real count which is updated since it is a separate

cache item.

This is a temp fix until post feed cache is updated correctly for comments.