revert back to using tag as a prop
This is needed because otherwise it is hard to convince
typescript to infer and use correct types for the Button component
or for it's forwardRef.

Previously I hacked it by providing a 3rd parameter to the Button
component function, but it was just temporary.
It is not a good solution in case React changes
the API in the future, this hack would break.