fix: scroll to top button position on iOS
Fixed by moving the ScrollToTop button to be part of the footer navbar
container instead. Also changed the positioning from fixed to absolute.
The navbar is already fixed positioned. This way, if the footer navbar
height or position changes, the button moves as well. So we should never
cover the scroll to top button with the navbar anymore.