ci: optimize ci time with some perks 😈 (#2865)
<img width="219" alt="image" src="https://github.com/dailydotdev/apps/assets/1993245/5e738ae8-b09d-451f-a76c-0ade3671b01b">

* I managed to reduce CI time by 73% to a ridicules time of 2.5min! ⚡ 
* Thanks to @rebelchris I learned about the ["Rerun failed tests"](https://circleci.com/docs/rerun-failed-tests/) feature. So you can rerun only failed tests now if needed 🤩 
* Parallel support in case we get more tests per project, we can easily spool up more workers per project

The essence of the change is that each package's CI now runs independently. In addition since the shared package lint takes time, I moved it to its own process to run it in parallel with the tests.

It's about time we get a better feedback loop. Enjoy!

eslint timing breakdown:
![image](https://github.com/dailydotdev/apps/assets/1993245/ee85d3ed-1b82-4451-b3df-c70fc0bbf5f9)