refactor: move pixels to _app (#4193)
Instead of adding them to specific pages, I rather have them everywhere in the webapp. We have some concerns about attribution so I'd like to see if this improves the situation.

The only caveat is hotjar. This means it will record all the sessions if we enable it. We should consider moving hotjar out and put it only where we need it. Right now it's disabled anyway