feat(funnels): checkout step
Introduce a checkout step which shows paddle checkout in fullscreen and transitions to next step on checkout.

Simple as it sounds, it required some changes to how we use paddle. For the storybook support I had to use history instead of router, since storybook doesn't have a router.

MI-843 #done