fix: leave funnel and initiate checkout
We now send leave funnel only on page unload and if the funnel wasn't completed. I also reworked some of the stuff because i don't like it when we disable exhaust deps. Usually it's a sign to use ref instead of callback/state.

Second fix is initiate checkout. Since we preload all pages, the checkout form is initialized as soon as the funnel starts which leads to initiate checkout event, I fixed that.